A bunch of core improvements (#288)
* wip

* progress

* messy, but passing messages

* ci: fix code style linting errors

* doc

* lift MessageReceiver

* test: Awareness propagation

* fix: persistWait not configurable, add onPersist test

* docs

* fix: Query awareness from other servers on load

* fix: Change import path (for now)

* refactor

* fix: Sync step propagation

* docs

* test: onDisconnect

* Add onCreatedDocument hook, reduce blocking

* fix build? Prefer src if available

* ci: fix code style linting errors

* docs: update navigation

* install dependencies

* update pubsub package.json, rename folder

* update imports to use package names

* move pubsub class to separate class

* refactor tests

* fix tests

* fix package-lock.json

* remove yarn lock file

* ci: fix code style linting errors

* update the ESLint config

* add some debugging output to check what happens in CI

* add more debugging output

* enable all other tests

* merge changes

* database: story the recent state, and only restore from the recent state

* rocksdb: revert change

* update dependencies

* fix eslint config

* refactor tests

* rewrite package-lock.json

* rewrite package-lock.json

* fix types

* ci: fix code style linting errors

* wip

* fix tests

* wip

* upgrade redis to 4.0

* update dependencies

* refactoring

* clean up

* replace old redis extension with the new approach

* ci: fix code style linting errors

* rename test context

* add new onStoreDocument hook

* add onStoreDocument to the logger extension

* make sure documents are removed from the memory after they are persisted, but only when they don’t have new connections

* skip a failing test

* debounce the onStoreDocument hook

* clean up

* wip

* update database and sqlite extensions to use the new onStoreDocument hook

* add two redis instances and two editors to the playground example

* fix status bar

* add onAwarenessUpdate hook

* redis: use unique ids instead of the hostname

* introduce new afterStoreDocument hook, rename onLoadedDocument to afterLoadDocument

* use global onStoreDocument hook

* refactoring

* make sure to only store document when changes came in through the WebSocket connection

* refactoring

* remove Redis specific onPersist hook

* fix tests

* add a priority to extensions

* server: add support for custom reply channels in the server MessageReceiver

* migrate the server/onStoreDocument test to AVA

* migrate more Redis tests to AVA

* refactor the tests

* migrate the onChange test

* provider: improve typings, make removeAllListeners public

* migrate the onAwarenessChange test

* ci: try to fix parcel build

* ci: remove dashboard build

* tests: minor refactorings

* clean up

* migrate more tests

* migrate more tests, add a few known failures for the redis extension

* increase the timeout for the redis/onChange test

* migrate even more tests

* restore broken test

* migrate another test

* server: allow to pass the configuration to the constructor

* disable workerThreads for AVA

* clean up the branch

+++ tests/server/onAwarenessUpdate.js
@@ -4,18 +4,31 @@ import WebSocket from 'ws'
 import { Hocuspocus } from '@hocuspocus/server'
 import { HocuspocusProvider } from '@hocuspocus/provider'
 
-let client
+let anotherClient
 const ydoc = new Y.Doc()
 
-context('server/onDestroy', () => {
-  it('has the server instance', done => {
+context('server/onAwarenessUpdate', () => {
+  it('onAwarenessUpdate hook is executed', done => {
+    let client
     const server = new Hocuspocus()
+
+    let called = false
 
     server.configure({
       port: 4000,
-      async onDestroy({ instance }) {
-        assert.strictEqual(instance, server)
+      onAwarenessUpdate: ({ states }) => {
+        if (called) {
+          return
+        }
+
+        called = true
+
+        server.destroy()
         client.destroy()
+
+        assert.strictEqual(states.length, 1)
+        assert.strictEqual(states[0].foo, 'bar')
+
         done()
       },
     }).listen()
@@ -26,7 +39,7 @@ context('server/onDestroy', () => {
       document: ydoc,
       WebSocketPolyfill: WebSocket,
       onConnect: () => {
-        server.destroy()
+        client.setAwarenessField('foo', 'bar')
       },
     })
   })
